+++ src/Exception/ErrorException.php
@@ -15,10 +15,8 @@ use function function_exists;
 /**
  * `ErrorException` represents a PHP error.
  * @psalm-type DebugBacktraceType = list<array{args?: array<mixed>, class?: class-string, file?: string, function?: string, line?: int, object?: object, type?: string}>
- *
- * @final
  */
-class ErrorException extends \ErrorException implements FriendlyExceptionInterface
+final class ErrorException extends \ErrorException implements FriendlyExceptionInterface
 {
     /** @psalm-suppress MissingClassConstType Private constants never change. */
     private const ERROR_NAMES = [