+++ tests/ErrorHandlerTest.php
@@ -6,7 +6,6 @@ namespace Yiisoft\ErrorHandler\Tests;
 
 use PHPUnit\Framework\Attributes\WithoutErrorHandler;
 use PHPUnit\Framework\TestCase;
-use Psr\Log\LoggerInterface;
 use RuntimeException;
 use Throwable;
 use Yiisoft\ErrorHandler\ErrorHandler;
@@ -16,14 +15,12 @@ use Yiisoft\ErrorHandler\ThrowableRendererInterface;
 final class ErrorHandlerTest extends TestCase
 {
     private ErrorHandler $errorHandler;
-    private LoggerInterface $loggerMock;
     private ThrowableRendererInterface $throwableRendererMock;
 
     protected function setUp(): void
     {
-        $this->loggerMock = $this->createMock(LoggerInterface::class);
         $this->throwableRendererMock = $this->createMock(ThrowableRendererInterface::class);
-        $this->errorHandler = new ErrorHandler($this->loggerMock, $this->throwableRendererMock);
+        $this->errorHandler = new ErrorHandler($this->throwableRendererMock);
         $this->errorHandler->memoryReserveSize(0);
     }